introduce base class for energy mailer
